refcator: separate runtime domain specific logic
